Created an Array, which saves Pac's perceptions
Array is called "sensEnv"--> Not really shure, if it saves everything
correctly. It also could save where Pac saw coins, so we could buid this
in our algorithm in order to let him "plan his path better". I already
got some ideas how we could create some code for doing this..
I also tried to call the sensor's perception in every round, but haven't
been very successful with my code, so i deleted it... We should talk
about that issue soon.
So maybe you could go over the code as it is at the moment and try to
check if everything is okay... Yesterday I felt like we could get a
problem with rounding the numbers we are working with...
But as I said before... We shoul talk to each other soon about those
topics...